% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEtimeKernCompute.R
\name{DEtimeKernCompute}
\alias{DEtimeKernCompute}
\title{Compute the DEtime kernel given the parameters and X}
\usage{
DEtimeKernCompute(kern, X, X2)
}
\arguments{
\item{kern}{DEtime kernel structure to be computed}

\item{X}{A two-column matrix where the first column of this matrix is the time points for control and perturbed conditions and the second column uses '1' to represent time points from control condition and '2' to represent time points from perturbed condition}

\item{X2}{Second input matrix to the kernel computation (forms the columns of the kernel)}
}
\value{
DEtime kernel structure.
}
\description{
Compute the DEtime kernel given the parameters and X.
}
\details{
\code{K <- DEtimeKernCompute(kern, X)} computes a DEtime kernel matrix given an input data matrix.
   \code{K <- DEtimeKernCompute(kern, X, X2)} computes a DEtime kernel matrix for the given kernel type given two input data matrices, one for the rows and one for the columns.
}
\examples{
kern <- list()
kern <- DEtimeKernParamInit(kern)
X <- matrix(c(seq(0,4),seq(0,4), rep(1,5),rep(2,5)),ncol=2)
K <- DEtimeKernCompute(kern, X)
}

