% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEtimeMatern12KernCompute.R
\name{DEtimeMatern12KernCompute}
\alias{DEtimeMatern12KernCompute}
\title{Compute the DEtime Matern12 kernel given the parameters and X}
\usage{
DEtimeMatern12KernCompute(kern, x, x2)
}
\arguments{
\item{kern}{DEtime Matern12 kernel structure to be computed}

\item{x}{Depending on the number of inputs, x can be the input data matrix (rows are data points) to the kernel computation, or the first input matrix to the kernel computation (forms the rows of the kernel)}

\item{x2}{Second input matrix to the kernel computation (forms the columns of the kernel)}
}
\value{
DEtime Matern12 kernel structure.
}
\description{
Compute the DEtime Matern12 kernel given the parameters and X.
}
\details{
\code{K <- DEtimeMatern12KernCompute(kern, x)} computes a DEtime Matern12 kernel matrix given an input data matrix.
   \code{K <- DEtimeMatern12KernCompute(kern, x1, x2)} computes a DEtime Matern12 kernel matrix for the given kernel type given two input data matrices, one for the rows and one for the columns.
}
\examples{
kern <- list()
kern <- DEtimeKernParamInit(kern)
K <- DEtimeMatern12KernCompute(kern, as.matrix(3:8))
}

