% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEtimeRBFKernCompute.R
\name{DEtimeRBFKernCompute}
\alias{DEtimeRBFKernCompute}
\title{Compute the DEtime RBF kernel given the parameters and X}
\usage{
DEtimeRBFKernCompute(kern, x, x2)
}
\arguments{
\item{kern}{DEtime RBF kernel structure to be computed}

\item{x}{Depending on the number of inputs, x can be the input data matrix (rows are data points) to the kernel computation, or the first input matrix to the kernel computation (forms the rows of the kernel)}

\item{x2}{Second input matrix to the kernel computation (forms the columns of the kernel)}
}
\value{
DEtime RBF kernel structure.
}
\description{
Compute the DEtime RBF kernel given the parameters and X.
}
\details{
\code{K <- DEtimeRBFKernCompute(kern, x)} computes a DEtime RBF kernel matrix given an input data matrix.
   \code{K <- DEtimeRBFKernCompute(kern, x1, x2)} computes a DEtime RBF kernel matrix for the given kernel type given two input data matrices, one for the rows and one for the columns.
}
\examples{
kern <- list()
kern <- DEtimeKernParamInit(kern)
K <- DEtimeRBFKernCompute(kern, as.matrix(3:8))
}

