% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealData.R
\name{RealData}
\alias{RealData}
\title{Experimental time-series data used for testing}
\format{A list which contains the time course data for 20 genes at time points [0, 2, 3, 4, 6, 7, 8, 10, 11, 12, 14, 16, 17.5] hrs with 4 replicates at each time point}
\description{
In this experiment, the aim is to study the transcriptional change occuring in Arabidopsis following inoculation with P. syringae pv. tomato DC3000 (PtoDC3000) versus the disarmed strain Pto DC3000hrpA. Two time series data are included: infection of Arabidopsis with virulent Pseudomonas syringage pv. tomato DC3000, which leads to disease developmen (perturbed condition 1) and infection of Arabidopsis with the disarmed strain DC3000hrpA (perturbed condition 2).
}
\details{
\itemize{
   \item ControlTimes Exprimental times for the perturbed condition 1
   \item ControlData Time series measurements for the perturbed condition 1
   \item PerturbedTimes Exprimental times for the perturbed condition 2
   \item PerturbedData Time series measurements for the perturbed condition 2
  }
}
\examples{
   data(RealData)
}

