% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatedData.R
\name{SimulatedData}
\alias{SimulatedData}
\title{Simulated time-series data used for testing}
\format{A list which contains simulated time course data for 5 genes.}
\description{
A simulated data contains two time-series, control and perturbed cases.
}
\details{
\itemize{
   \item ControlTimes Exprimental times for simulated control condition
   \item ControlData Time series data for simulated control condition
   \item PerturbedTimes Exprimental times for simulated perturbed condition
   \item PerturbedData Time series data for simulated perturbed condition
  }
}
\examples{
   data(SimulatedData)
}

