% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEtime_rank.R
\name{DEtime_rank}
\alias{DEtime_rank}
\title{Calculating the log-likelihood ratio of the biological time course data}
\usage{
DEtime_rank(times, ControlData, PerturbedData, gene_ID = NULL,
  savefile = TRUE)
}
\arguments{
\item{times}{Experimental time point at which time course biological data are measured}

\item{ControlData}{Time course data measured under control condition}

\item{PerturbedData}{Time course data measured under perturbed condition}

\item{gene_ID}{ID of these genes addressed in this study}

\item{savefile}{A BOOLEAN argument which is used to indicate if the ranking list will be saved or not}
}
\value{
\code{DEtime_rank} returns a dataframe object whose first column is the gene_ID and second column is the Loglikelihood_ratio of the named gene.
}
\description{
\code{DEtime_rank} intends to rank biological time course data measured under control and perturbed conditions. In the function, an independent GP model and a combined GP model are used to model the data separately, the difference between the log-likelihood is used as the rank factor for the dataset. A high rank factor normally indicates better differential expression.
}
\details{
Both control and perturbed data have to be measured at the same time points with the same number of replicates. Replicates are required to be obtained across all time points. ControlData and PerturbedData are two matrices where each row represents the time course data for one particular gene under either control or perturbed condition. The columns for both ControlData and  PerturbedData are ordered by the time sequencing followed by replicates.
}
\examples{
### import simulated data
data(SimulatedData)
### calculating loglikelihood ratio
res <- DEtime_rank(times = times, ControlData = ControlData, 
  PerturbedData=PerturbedData, gene_ID=gene_ID)
}

