% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DEtimeKernDiagCompute.R
\name{DEtimeKernDiagCompute}
\alias{DEtimeKernDiagCompute}
\title{Retrieve the diagonal of the DEtime kernel}
\usage{
DEtimeKernDiagCompute(kern, x)
}
\arguments{
\item{kern}{DEtime kernel structure to be computed}

\item{x}{Depending on the number of inputs, x can be the input data matrix (rows are data points) to the kernel computation, or the first input matrix to the kernel computation (forms the rows of the kernel)}
}
\value{
Kd Vector containing computed diagonal elements of the kernel structure
}
\description{
Retrieve the diagonal of the DEtime kernel matrix.
}
\details{
\code{Kd <- DEtimekernDiagCompute(kern, x)} computes the diagonal of a DEtime kernel matrix for the given kernel.
}
\examples{
kern <- list()
kern <- DEtimeKernParamInit(kern)
K <- DEtimeKernCompute(kern, as.matrix(3:8))
Kd <- DEtimeKernDiagCompute(kern, as.matrix(3:8))
}

