% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEtimeKernDiagCompute.R
\name{DEtimeKernDiagCompute}
\alias{DEtimeKernDiagCompute}
\title{Retrieve the diagonal of the DEtime kernel}
\usage{
DEtimeKernDiagCompute(kern, X)
}
\arguments{
\item{kern}{DEtime kernel structure to be computed}

\item{X}{A two-column matrix where the first column of this matrix is the time points for control and perturbed conditions and the second column uses '1' to represent time points from control condition and '2' to represent time points from perturbed condition}
}
\value{
Kd Vector containing computed diagonal elements of the kernel structure
}
\description{
Retrieve the diagonal of the DEtime kernel matrix.
}
\details{
\code{Kd <- DEtimekernDiagCompute(kern, X)} computes the diagonal of a DEtime kernel matrix for the given kernel.
}
\examples{
kern <- list()
kern <- DEtimeKernParamInit(kern)
X <- matrix(c(seq(3:8),seq(4:8),rep(1,6),rep(2,5)),ncol=2)
Kd <- DEtimeKernDiagCompute(kern, X)
}

