% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEtimeKernExpandParam.R
\name{DEtimeKernExpandParam}
\alias{DEtimeKernExpandParam}
\title{Update a DEtime kernel structure with new parameters}
\usage{
DEtimeKernExpandParam(kern, params)
}
\arguments{
\item{kern}{The DEtime kernel structure whose parameters to be expanded}

\item{params}{Vector of parameters}
}
\value{
kern Updated DEtime kernel structure
}
\description{
Update a DEtime kernel structure with new parameters.
}
\details{
\code{DEtimeKernExpandParam} returns a DEtime kernel structure filled with the parameters in the given vector. This is used as a helper function to enable parameters to be optimised in, for example, the optimisation functions.
}
\examples{
kern <- list()
kern <- DEtimeKernParamInit(kern)
X <- matrix(c(seq(0,4),seq(0,4), rep(1,5),rep(2,5)),ncol=2)
K <- DEtimeKernCompute(kern, X)
### the inverseWidth, variance and testing perturbation point for DEtime Kernel function
params <- c(2, 1, 20)  
Knew <- DEtimeKernExpandParam(K, params)
}
\seealso{
\code{\link{DEtimeKernExtractParam}}
}

