% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEtimeKernExtractParam.R
\name{DEtimeKernExtractParam}
\alias{DEtimeKernExtractParam}
\title{Extract the parameters used in a DEtime kernel}
\usage{
DEtimeKernExtractParam(kern, only.values = TRUE,
  untransformed.values = TRUE)
}
\arguments{
\item{kern}{The DEtime kernel structure whose parameters to be extracted from}

\item{only.values}{A BOOLEAN parameter with value TRUE or FALSE. When it is TRUE, only the values of the paramter vector are returned, when it is FALSE, both the names and the values of the paramter vector are returned}

\item{untransformed.values}{A BOOLEAN parameter with value TRUE or FALSE. When set to TRUE, the untransformed values of the parameter vector are returned}
}
\value{
params The parameters used in the DEtime kernel
}
\description{
Extract the parameters from a DEtime kernel.
}
\details{
\code{DEtimeKernExtractParam} function returns the hyperparameters used in a DEtime kernel structure.
}
\examples{
kern <- list()
kern <- DEtimeKernParamInit(kern)
params <- DEtimeKernExtractParam(kern)
}
\seealso{
\code{\link{DEtimeKernExpandParam}}
}

