% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DEtimeKernParamInit.R
\name{DEtimeKernParamInit}
\alias{DEtimeKernParamInit}
\title{DEtime kernel parameter initialisation}
\usage{
DEtimeKernParamInit(kern)
}
\arguments{
\item{kern}{The DEtime kernel structure of which the parameters will be initialised}
}
\value{
kern The DEtime kernel structure with the parameters initialised
}
\description{
initialises the parameters of a DEtime kernel
}
\examples{
kern <- list()
K <- DEtimeKernParamInit(kern)
}
\seealso{
\code{\link{DEtimeKernCompute}}
}

