% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DEtime.R
\name{plot_DEtime}
\alias{plot_DEtime}
\title{Plot the results from DEtime function}
\usage{
plot_DEtime(DEtimeOutput, BestPerturbPos = NULL, plot_gene_ID = NULL)
}
\arguments{
\item{DEtimeOutput}{The returned value from \code{\link{DEtime_infer}} function}

\item{BestPerturbPos}{The statistical factor used for the optimal estimate to the perturbation point. You can choose "mean", "median" or "MAP". The corresponding statistical inference of the posterior distribution of the pertubation points will be adopted as the best estimate to the perturbation. "MAP" will be used if this parameter is not provided.}

\item{plot_gene_ID}{The gene_IDs of those genes whose GP regression and posterior distribution of the perturbation time points will be plotted. If not supplied, all the genes will be plotted.}
}
\value{
A figure for each gene whhich illustrates the GP regression (lower panel) as well as posterior distribution of its perturbation points (upper panel).
}
\description{
This function plots the results returned from \code{\link{DEtime_infer}} function. The produced figures show the the posterior distribution of inferred perturbation time points on the upper panel and Gaussian Regression of the original data on the lower panel.
}
\examples{
data(SimulatedData)
res <- DEtime_infer(ControlTimes = ControlTimes, ControlData = ControlData, 
       PerturbedTimes = PerturbedTimes, PerturbedData = PerturbedData)
plot_DEtime(res,BestPerturbPos="mean", plot_gene_ID=c('1','2'))
}
\seealso{
\code{\link{DEtime_infer}} \code{\link{print_DEtime}}
}

