% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RealData.R
\name{RealData}
\alias{RealData}
\title{Experimental time-series data used for testing}
\format{A list which contains the time course data for 20 genes at time points [0, 2, 3, 4, 6, 7, 8, 10, 11, 12, 14, 16, 17.5] hrs with 4 replicates at each time point}
\description{
In this experiment, the aim is to study the transcriptional change occuring in Arabidopsis following inoculation with P. syringae pv. tomato DC3000 (PtoDC3000) versus the disarmed strain Pto DC3000hrpA. Three time series data are included: a control time series comprising challenge with 10 mM MgCl2 (control condition), infection of Arabidopsis with virulent Pseudomonas syringage pv. tomato DC3000, which leads to disease developmen (perturbed condition 1) and infection of Arabidopsis with the disarmed strain DC3000hrpA (perturbed condition 2).
}
\details{
\itemize{
   \item times Simulated exprimental times
   \item times_test Predefined perturbation time points
   \item ControlData Time series data for control condition
   \item PerturbedData Time series data for perturbed condition
   \item replicate_no Number of replicates used in the simulation
   \item gene_no Number of genes involved in the study
   \item gene_ID ID for studied genes
  }
}
\examples{
data(RealData)
}

